[fix] member-service에서 security 관련 코드 삭제
