+++ member-service/src/main/java/com/rainbowgon/memberservice/domain/member/controller/MemberController.java
@@ -9,9 +9,9 @@ import com.rainbowgon.memberservice.domain.member.dto.response.oauth.OAuthProfil
 import com.rainbowgon.memberservice.domain.member.service.KakaoLoginService;
 import com.rainbowgon.memberservice.domain.member.service.MemberService;
 import com.rainbowgon.memberservice.global.client.NotificationServiceClient;
+import com.rainbowgon.memberservice.global.jwt.dto.JwtTokenDto;
 import com.rainbowgon.memberservice.global.response.JsonResponse;
 import com.rainbowgon.memberservice.global.response.ResponseWrapper;
-import com.rainbowgon.memberservice.global.security.dto.JwtTokenDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -43,7 +43,7 @@ public class MemberController {
 
         return ResponseEntity.ok(OAuthProfileResDto.fromKakao(kakaoProfileResDto));
     }
-    
+
 
     /**
      * 회원가입
