+++ member-service/src/main/java/com/rainbowgon/memberservice/domain/member/service/MemberService.java
@@ -4,7 +4,7 @@ import com.rainbowgon.memberservice.domain.member.dto.request.MemberCreateReqDto
 import com.rainbowgon.memberservice.domain.member.dto.request.MemberPhoneReqDto;
 import com.rainbowgon.memberservice.domain.member.dto.request.MemberUpdateReqDto;
 import com.rainbowgon.memberservice.domain.member.dto.response.MemberInfoResDto;
-import com.rainbowgon.memberservice.global.security.dto.JwtTokenDto;
+import com.rainbowgon.memberservice.global.jwt.dto.JwtTokenDto;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.UUID;
