+++ member-service/src/main/java/com/rainbowgon/memberservice/domain/profile/service/ProfileServiceImpl.java
@@ -9,8 +9,8 @@ import com.rainbowgon.memberservice.domain.profile.repository.ProfileRepository;
 import com.rainbowgon.memberservice.global.entity.NotificationStatus;
 import com.rainbowgon.memberservice.global.error.exception.ProfileNotFoundException;
 import com.rainbowgon.memberservice.global.error.exception.ProfileUnauthorizedException;
-import com.rainbowgon.memberservice.global.security.JwtTokenProvider;
-import com.rainbowgon.memberservice.global.security.dto.JwtTokenDto;
+import com.rainbowgon.memberservice.global.jwt.JwtTokenProvider;
+import com.rainbowgon.memberservice.global.jwt.dto.JwtTokenDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
