--- member-service/src/main/java/com/rainbowgon/memberservice/global/config/SecurityConfigure.java
@@ -1,49 +1,0 @@
-package com.rainbowgon.memberservice.global.config;
-
-import com.rainbowgon.memberservice.global.security.JwtTokenProvider;
-import com.rainbowgon.memberservice.global.security.filter.CustomExceptionFilter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class SecurityConfigure {
-
-    private final JwtTokenProvider jwtTokenProvider;
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception {
-
-        httpSecurity
-                .cors()
-                .and()
-                .csrf().disable()
-
-                .sessionManagement()
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-                .and()
-                .authorizeRequests()
-                .antMatchers("/members/signup/**").permitAll()
-                .antMatchers("/members/login/**").permitAll()
-                .antMatchers("/reviews/themes/one/**").permitAll()
-                .antMatchers("/members/phone/**").permitAll()
-                .antMatchers(("/actuator/**")).permitAll()
-                .antMatchers("/clients/**").permitAll() // 서버 간 통신
-                .antMatchers("/members/test/**").permitAll() // 서버 간 통신 테스트
-                .anyRequest().authenticated();
-
-        httpSecurity
-                .addFilterBefore(new CustomExceptionFilter(), UsernamePasswordAuthenticationFilter.class);
-
-        return httpSecurity.build();
-    }
-
-
-}
